+++ jeap-deploymentlog-domain/src/main/java/ch/admin/bit/jeap/deploymentlog/domain/DeploymentTerminalMetricEvent.java
@@ -1,6 +1,7 @@
 package ch.admin.bit.jeap.deploymentlog.domain;
 
 import java.time.ZonedDateTime;
+import java.util.Set;
 import java.util.UUID;
 
 public record DeploymentTerminalMetricEvent(
@@ -9,6 +10,7 @@ public record DeploymentTerminalMetricEvent(
         String system,
         String component,
         String environment,
+        Set<DeploymentType> deploymentTypes,
         DeploymentState state,
         ZonedDateTime startedAt,
         ZonedDateTime endedAt) {
@@ -21,6 +23,7 @@ public record DeploymentTerminalMetricEvent(
                 component.getSystem().getName(),
                 component.getName(),
                 deployment.getEnvironment().getName(),
+                Set.copyOf(deployment.getDeploymentTypes()),
                 deployment.getState(),
                 deployment.getStartedAt(),
                 deployment.getEndedAt());